+++ account/accesskey/create_request.go
@@ -12,27 +12,16 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package accountkey
+package accesskey
 
 import (
-	"context"
-
-	objectstorage "github.com/sacloud/object-storage-api-go"
-	v1 "github.com/sacloud/object-storage-api-go/apis/v1"
+	"github.com/sacloud/packages-go/validate"
 )
 
-func (s *Service) Find(req *FindRequest) ([]*v1.AccountKey, error) {
-	return s.FindWithContext(context.Background(), req)
+type CreateRequest struct {
+	SiteId string `service:"-" validate:"required"`
 }
 
-func (s *Service) FindWithContext(ctx context.Context, req *FindRequest) ([]*v1.AccountKey, error) {
-	if req == nil {
-		req = &FindRequest{}
-	}
-	if err := req.Validate(); err != nil {
-		return nil, err
-	}
-
-	client := objectstorage.NewAccountOp(s.client)
-	return client.ListAccessKeys(ctx, req.SiteId)
+func (req *CreateRequest) Validate() error {
+	return validate.New().Struct(req)
 }
