+++ account/accesskey/find_request.go
@@ -12,17 +12,14 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package accountkey
+package accesskey
 
-import (
-	"github.com/sacloud/packages-go/validate"
-)
+import "github.com/sacloud/packages-go/validate"
 
-type DeleteRequest struct {
+type FindRequest struct {
 	SiteId string `service:"-" validate:"required"`
-	Id     string `service:"-" validate:"required"` // リソースId
 }
 
-func (req *DeleteRequest) Validate() error {
+func (req *FindRequest) Validate() error {
 	return validate.New().Struct(req)
 }
