+++ account/accesskey/read_request.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package accountkey
+package accesskey
 
 import (
 	"github.com/sacloud/packages-go/validate"
